Simplify / future-proof the NumPy implementation of the weighting classes.
The previous version made hard-coded decisions of low-level operations. These really are specific to NumPy
and the various computational packages. This allowed getting fairly good performance on CPUs.

In the future, NumPy will be mostly a fallback; performance-critical applications should use GPU-enabled
storage backends instead. This means it is less critical to optimize CPU performance for NumPy, but more
important to keep things general.
This commit proposes using only NumPy functions that are in the Python Array API for the weighting classes,
meaning they can easily be generalised to e.g. PyTorch.